feat: enable provider plugin cache
By default, `terraform init` downloads plugins into a subdirectory
of the working directory so that each working directory is self-contained.

But enable the plugin cache,
then allows each distinct plugin binary to be downloaded only once.

This contributes to speeding up the `terraform init` command
when it is executed many times in a workflow.

See details:

- https://developer.hashicorp.com/terraform/cli/config/config-file#provider-plugin-cache