feat: add support for installing tfcmt
tfcmt is a CLI tool to improve the experience of CI of Terraform.
By posting the result of terraform plan and terraform apply to GitHub Pull Requests as a comment,
you can know the result quickly without browsing the CI web page.

- https://github.com/suzuki-shunsuke/tfcmt

It is convenient to use tfcmt with Terraform in GitHub Actions.
Therefore, we will also install tfcmt in this action.